refactor: split server route ownership [audit:Q-08]
