+++ src/server/http.ts
@@ -207,6 +207,40 @@ export async function parseBody(
   }
 }
 
+export function isJsonObject(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+
+export async function parseObjectBody(
+  req: IncomingMessage,
+  res: ServerResponse,
+  options: ParseBodyOptions = {},
+): Promise<Record<string, unknown> | null> {
+  const body = await parseBody(req, res, options);
+  if (body === undefined) return null;
+  if (!isJsonObject(body)) {
+    json(
+      res,
+      options.invalidResponse?.envelope ?? { error: "JSON body must be an object" },
+      options.invalidResponse?.status ?? 400,
+    );
+    return null;
+  }
+  return body;
+}
+
+export function hasOptionalType(
+  body: Record<string, unknown>,
+  key: string,
+  type: "string" | "number" | "boolean",
+): boolean {
+  return body[key] === undefined || typeof body[key] === type;
+}
+
+export function hasOnlyKeys(body: Record<string, unknown>, allowedKeys: ReadonlySet<string>): boolean {
+  return Object.keys(body).every(key => allowedKeys.has(key));
+}
+
 /** Generate a cryptographically random base64 nonce for CSP. */
 export function generateCspNonce(): string {
   return randomBytes(16).toString("base64");
